requirements.txt
aiofiles==23.2.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
fastapi==0.112.2
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gradio==4.42.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_resources==6.4.4
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
Nuitka==2.6.9
numpy==2.1.0
openai==1.43.0
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.1
requests==2.32.3
rich==13.8.0
ruff==0.6.3
semantic-version==2.10.0
setuptools==78.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
zstandard==0.23.0
